requirements.txt
SQLAlchemy>=1.0.8
sqlalchemy-migrate>=0.10.0
black
msal>=1.4.1
python-dateutil>=2.8.1
tabulate>=0.8.7